Make curl trait-based
