+++ curl/src/lib.rs
@@ -2,6 +2,6 @@ extern crate trytes;
 
 pub mod constants;
 pub mod curl;
+pub mod simple;
 
-pub use constants::*;
 pub use curl::*;
